refactor: BookCartContext를 BagContext로 리네임
