+++ components/templates/BookPrevewListTemplate.tsx
@@ -2,7 +2,7 @@
 import { BookPreview } from "@/types/BookPreview";
 import { EmptySearchResult } from "../molecules/EmptySearchResult";
 import { SearchableBookInfo } from "../organisms/book/preview/SearchableBookInfo";
-import { useBookCart } from "@/contexts/BookCartContext";
+import { useBookCart } from "@/contexts/BagContext";
 import { useState } from "react";
 import { InfoToast } from "../molecules/toast/InfoToast";
